+++ src/gwt/src/org/rstudio/studio/client/workbench/views/files/FilesConstants.java
@@ -113,4 +113,5 @@ public interface FilesConstants extends com.google.gwt.i18n.client.Messages {
     String nameHeaderText();
     String symbolicLinkBadgeLabel();
     String aliasBadgeLabel();
+    String shortcutBadgeLabel();
 }